requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.110.0
fastapi-utils==0.8.0
h11==0.16.0
httptools==0.7.1
idna==3.11
mongoengine==0.29.1
mypy_extensions==1.1.0
passlib==1.7.4
psutil==5.9.8
pyasn1==0.6.1
pycparser==2.23
pydantic==2.8.2
pydantic-settings==2.2.0
pydantic_core==2.20.1
pymongo==4.15.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
typing-inspect==0.9.0
typing_extensions==4.15.0
uvicorn==0.29.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
